% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcClassify.R, R/bcClassify1.R
\docType{package}
\name{BLCAclassifier-package}
\alias{BLCAclassifier-package}
\title{BLCAclassifier Package}
\description{
Bladder Cancer Immune Subtype Classification

Bladder Cancer Immune Subtype Classification
}
\section{Usage}{

\preformatted{
library(BLCAclassifier)
results <- bcClassify(expression_matrix)
}


\preformatted{
library(BLCAclassifier)
results <- bcClassify(expression_matrix)
}
}

\author{
\strong{Maintainer}: GD Chu \email{chugd266000@126.com}

}
\keyword{internal}
