% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcClassify.R, R/bcClassify1.R
\name{load_ntp_template}
\alias{load_ntp_template}
\title{Load NTP Template}
\usage{
load_ntp_template()

load_ntp_template()
}
\value{
Template data frame with probe and class columns

Template data frame with probe and class columns
}
\description{
Internal function to load NTP template from package data

Internal function to load NTP template from package data
}
