% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcClassify.R, R/bcClassify1.R
\name{ntp_classify}
\alias{ntp_classify}
\title{NTP Classification Algorithm (Bug-Fixed)}
\usage{
ntp_classify(expr_data, template_df)

ntp_classify(expr_data, template_df)
}
\arguments{
\item{expr_data}{Standardized expression matrix}

\item{template_df}{Template data frame}
}
\value{
Vector of predicted subtypes

Vector of predicted subtypes
}
\description{
Internal function for gene signature-based classification

Internal function for gene signature-based classification
}
